fix(logging): annotate the printf-like entry points, and fix what the checking found (#216) (#359)
dbg_printf, __syslog and the printf/sprintf/snprintf/fprintf/v*
family now carry __attribute__((format(printf, ...))), so every
format/argument mismatch in the tree is a compile-time error under
the existing -Werror.

The annotation found four real bugs, live at runtime:

- vfs_mkdir/rmdir/creat/stat/symlink logged the superblock failure
  with a %s and no argument, printing stack contents as the path.
- ftok's stat failure printed %s with no argument, same class.
- vfs_readlink logged a size_t bufsize through %d.
- shell's SIGCHLD sigaction failure logged the signal number
  through %s, and its .shellrc message had a conversion too many.

The rest of the ~120 diagnostics are type drift between specifier
and argument: size_t/ssize_t/long now printed with %lu/%ld, int-sized
values with %d/%u, typed pointers passed to %p cast to (void *)
(through an integer for the function pointers, which -Wpedantic
refuses to convert directly), and PRINT_REGS's %-0N flags cleaned of
the '0' that '-' already ignored.

Verified with GCC 13 and Clang 18 under the project's warning set.

Fixes #216.

+++ kernel/src/devices/pci.c
@@ -1179,7 +1179,7 @@ void pci_debug_scan(void)
     pr_default("Counting PCI entities...\n");
     size_t count = 0;
     pci_scan(&__scan_count, -1, &count);
-    pr_default("Total PCI entities: %d\n", count);
+    pr_default("Total PCI entities: %u\n", (unsigned int)count);
 
     pr_default("Scanning PCI entities...\n");
     pci_scan(&__scan_hit_list, -1, NULL);
